requirements.txt
msgspec
numpy
opencv-python
black
pillow
tqdm